Changed the IRIS detection code and made it angle based which makes it more accurate.
Added a faceInCenter which makes sure that a face in present right
infront of the camera and that the face is not at an extreme angle
facing left or right. Only if the face is present does the face
vectoring start to save processing power and false alarms

Created a seperate folder of the main file and supporting
functions for the Raspberry pi itself. The speaker and GPIO based
functions are different for a Raspberry pi and a laptop